% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_best.R
\name{plot_best}
\alias{plot_best}
\title{Plots all varibles at once in best step according to \code{criterion}}
\usage{
plot_best(object)
}
\arguments{
\item{object}{[\code{GAMBoost}]\cr including \code{y}
# @param title [\code{character(1)}]\cr Change title. Default "object: Best AIC at Iteration x"}
}
\description{
Plots all varibles at once in best step according to \code{criterion}
}
\author{
Philipp J Roesch
}

