% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/freq_selected.R
\name{plot_freq}
\alias{plot_freq}
\title{Plots frequency of selected variables at specified step}
\usage{
plot_freq(object, at.step = NULL)
}
\arguments{
\item{object}{[\code{GAMBoost}]\cr including \code{y}}

\item{at.step}{[\code{numeric(1)}]\cr specifies iteration step}
}
\description{
xx
}
\author{
Philipp J Roesch
}

