% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ICs.R
\name{plot_ICs}
\alias{plot_ICs}
\title{Plot information criterion for all iterations}
\usage{
plot_ICs(object)
}
\arguments{
\item{object}{[\code{GAMBoost}]\cr including \code{y}
# @param title title [\code{character(1)}]\cr Change title.}
}
\description{
Plots Deviance and AIC in one plot and indicates the iteration with the best AIC
}
\author{
Philipp J Roesch
}

