% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ploter.R
\name{ploter}
\alias{ploter}
\title{Plots GAMBoost objects with ggplot2}
\usage{
ploter(object, select = NULL, at.step = NULL, print = TRUE)
}
\arguments{
\item{object}{[\code{GAMBoost}]\cr including \code{y}}

\item{select}{[\code{vector(1)}]\cr Indicating the variables which should be plotted.}

\item{at.step}{[\code{integer(1)}]\cr Numeric value indicating which iteration step of GAMBoost is used.}

\item{print}{[\code{logical(1)}]\cr Sepcifies if should be printed or not. Default is TRUE.}
}
\description{
Same functionality (and almost same code) like in the orginal
FIX_ME_PLEASE: If several plots are created (which is default)
only the last plot is returned
NOTE: phi does not need to be specified anymore
\code{plot.GAMBoost} function but done with ggplot2.
Not tested
}
\author{
Philipp J Roesch
}

